# potion-seller/openapi.yaml
openapi: 3.0.0
info:
  title: Potion Shop Inventory API
  version: 1.0.0
  description: Simple API for listing and buying magical potions.

servers:
  - url: https://api.magic-shop.example.com

paths:
  /items:
    get:
      summary: List magical items
      description: Returns a paginated list of magical items in the shop.
      parameters:
        - name: limit
          in: query
          description: Maximum number of items to return.
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          description: Number of items to skip before starting to collect the result set.
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: List of magical items.
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Item'
                  total:
                    type: integer
                    description: Total number of items available.
                  limit:
                    type: integer
                  offset:
                    type: integer

  /items/{itemId}:
    get:
      summary: Get a magical item
      description: Returns details for a single magical item, including stock level.
      parameters:
        - name: itemId
          in: path
          required: true
          description: Unique identifier of the item.
          schema:
            type: string
      responses:
        '200':
          description: Magical item details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        '404':
          description: Item not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /items/{itemId}/stock:
    get:
      summary: Get remaining stock for an item
      description: Returns how many units of the given magical item are left.
      parameters:
        - name: itemId
          in: path
          required: true
          description: Unique identifier of the item.
          schema:
            type: string
      responses:
        '200':
          description: Remaining stock for the item.
          content:
            application/json:
              schema:
                type: object
                properties:
                  itemId:
                    type: string
                  remaining:
                    type: integer
                    description: Units left in stock.
        '404':
          description: Item not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /items/{itemId}/purchase:
    post:
      summary: Buy a magical item
      description: Attempts to purchase one or more units of an item and returns the updated stock level.
      parameters:
        - name: itemId
          in: path
          required: true
          description: Unique identifier of the item.
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - quantity
              properties:
                quantity:
                  type: integer
                  minimum: 1
                  description: Number of units to purchase.
      responses:
        '200':
          description: Purchase successful.
          content:
            application/json:
              schema:
                type: object
                properties:
                  itemId:
                    type: string
                  purchased:
                    type: integer
                  remaining:
                    type: integer
                    description: Units left in stock after the purchase.
        '400':
          description: Invalid request (e.g. quantity too large or item out of stock).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Item not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  schemas:
    Item:
      type: object
      required:
        - name
        - price
        - stock
      properties:
        name:
          type: string
          description: Display name of the item.
          example: Elder Wand
        description:
          type: string
          description: Short description of the item.
          example: A legendary wand said to be unbeatable in duels.
        price:
          type: number
          format: float
          description: Price per unit.
          example: 249.99
        stock:
          type: integer
          description: Units currently in stock.
          example: 3
        rarity:
          type: string
          description: Rarity classification of the item.
          enum: [common, uncommon, rare, epic, legendary]
          example: legendary

    Error:
      type: object
      properties:
        code:
          type: string
          example: OUT_OF_STOCK
        message:
          type: string
          example: Not enough items in stock to fulfill this purchase.