# potion-seller/src/api/gen/models.go
// Package gen provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.1 DO NOT EDIT.
package gen

// Defines values for ItemRarity.
const (
	Common    ItemRarity = "common"
	Epic      ItemRarity = "epic"
	Legendary ItemRarity = "legendary"
	Rare      ItemRarity = "rare"
	Uncommon  ItemRarity = "uncommon"
)

// Error defines model for Error.
type Error struct {
	Code    *string `json:"code,omitempty"`
	Message *string `json:"message,omitempty"`
}

// Item defines model for Item.
type Item struct {
	// Description Short description of the item.
	Description *string `json:"description,omitempty"`

	// Name Display name of the item.
	Name string `json:"name"`

	// Price Price per unit.
	Price float32 `json:"price"`

	// Rarity Rarity classification of the item.
	Rarity *ItemRarity `json:"rarity,omitempty"`

	// Stock Units currently in stock.
	Stock int `json:"stock"`
}

// ItemRarity Rarity classification of the item.
type ItemRarity string

// GetItemsParams defines parameters for GetItems.
type GetItemsParams struct {
	// Limit Maximum number of items to return.
	Limit *int `form:"limit,omitempty" json:"limit,omitempty"`

	// Offset Number of items to skip before starting to collect the result set.
	Offset *int `form:"offset,omitempty" json:"offset,omitempty"`
}

// PostItemsItemIdPurchaseJSONBody defines parameters for PostItemsItemIdPurchase.
type PostItemsItemIdPurchaseJSONBody struct {
	// Quantity Number of units to purchase.
	Quantity int `json:"quantity"`
}

// PostItemsItemIdPurchaseJSONRequestBody defines body for PostItemsItemIdPurchase for application/json ContentType.
type PostItemsItemIdPurchaseJSONRequestBody PostItemsItemIdPurchaseJSONBody