# potion-seller/src/api/gen/routes.go
// Package gen provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.1 DO NOT EDIT.
package gen

import (
	"context"
	"encoding/json"
	"fmt"
	"net/http"

	"github.com/gin-gonic/gin"
	"github.com/oapi-codegen/runtime"
	strictgin "github.com/oapi-codegen/runtime/strictmiddleware/gin"
)

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// List magical items
	// (GET /items)
	GetItems(c *gin.Context, params GetItemsParams)
	// Get a magical item
	// (GET /items/{itemId})
	GetItemsItemId(c *gin.Context, itemId string)
	// Buy a magical item
	// (POST /items/{itemId}/purchase)
	PostItemsItemIdPurchase(c *gin.Context, itemId string)
	// Get remaining stock for an item
	// (GET /items/{itemId}/stock)
	GetItemsItemIdStock(c *gin.Context, itemId string)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandler       func(*gin.Context, error, int)
}

type MiddlewareFunc func(c *gin.Context)

// GetItems operation middleware
func (siw *ServerInterfaceWrapper) GetItems(c *gin.Context) {
	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params GetItemsParams

	// ------------- Optional query parameter "limit" -------------

	err = runtime.BindQueryParameter("form", true, false, "limit", c.Request.URL.Query(), &params.Limit)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter limit: %w", err), http.StatusBadRequest)
		return
	}

	// ------------- Optional query parameter "offset" -------------

	err = runtime.BindQueryParameter("form", true, false, "offset", c.Request.URL.Query(), &params.Offset)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter offset: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.GetItems(c, params)
}

// GetItemsItemId operation middleware
func (siw *ServerInterfaceWrapper) GetItemsItemId(c *gin.Context) {
	var err error

	// ------------- Path parameter "itemId" -------------
	var itemId string

	err = runtime.BindStyledParameterWithOptions("simple", "itemId", c.Param("itemId"), &itemId, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter itemId: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.GetItemsItemId(c, itemId)
}

// PostItemsItemIdPurchase operation middleware
func (siw *ServerInterfaceWrapper) PostItemsItemIdPurchase(c *gin.Context) {
	var err error

	// ------------- Path parameter "itemId" -------------
	var itemId string

	err = runtime.BindStyledParameterWithOptions("simple", "itemId", c.Param("itemId"), &itemId, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter itemId: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.PostItemsItemIdPurchase(c, itemId)
}

// GetItemsItemIdStock operation middleware
func (siw *ServerInterfaceWrapper) GetItemsItemIdStock(c *gin.Context) {
	var err error

	// ------------- Path parameter "itemId" -------------
	var itemId string

	err = runtime.BindStyledParameterWithOptions("simple", "itemId", c.Param("itemId"), &itemId, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter itemId: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.GetItemsItemIdStock(c, itemId)
}

// GinServerOptions provides options for the Gin server.
type GinServerOptions struct {
	BaseURL      string
	Middlewares  []MiddlewareFunc
	ErrorHandler func(*gin.Context, error, int)
}

// RegisterHandlers creates http.Handler with routing matching OpenAPI spec.
func RegisterHandlers(router gin.IRouter, si ServerInterface) {
	RegisterHandlersWithOptions(router, si, GinServerOptions{})
}

// RegisterHandlersWithOptions creates http.Handler with additional options
func RegisterHandlersWithOptions(router gin.IRouter, si ServerInterface, options GinServerOptions) {
	errorHandler := options.ErrorHandler
	if errorHandler == nil {
		errorHandler = func(c *gin.Context, err error, statusCode int) {
			c.JSON(statusCode, gin.H{"msg": err.Error()})
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandler:       errorHandler,
	}

	router.GET(options.BaseURL+"/items", wrapper.GetItems)
	router.GET(options.BaseURL+"/items/:itemId", wrapper.GetItemsItemId)
	router.POST(options.BaseURL+"/items/:itemId/purchase", wrapper.PostItemsItemIdPurchase)
	router.GET(options.BaseURL+"/items/:itemId/stock", wrapper.GetItemsItemIdStock)
}

type GetItemsRequestObject struct {
	Params GetItemsParams
}

type GetItemsResponseObject interface {
	VisitGetItemsResponse(w http.ResponseWriter) error
}

type GetItems200JSONResponse struct {
	Items  *[]Item `json:"items,omitempty"`
	Limit  *int    `json:"limit,omitempty"`
	Offset *int    `json:"offset,omitempty"`

	// Total Total number of items available.
	Total *int `json:"total,omitempty"`
}

func (response GetItems200JSONResponse) VisitGetItemsResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type GetItemsItemIdRequestObject struct {
	ItemId string `json:"itemId"`
}

type GetItemsItemIdResponseObject interface {
	VisitGetItemsItemIdResponse(w http.ResponseWriter) error
}

type GetItemsItemId200JSONResponse Item

func (response GetItemsItemId200JSONResponse) VisitGetItemsItemIdResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type GetItemsItemId404JSONResponse Error

func (response GetItemsItemId404JSONResponse) VisitGetItemsItemIdResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(404)

	return json.NewEncoder(w).Encode(response)
}

type PostItemsItemIdPurchaseRequestObject struct {
	ItemId string `json:"itemId"`
	Body   *PostItemsItemIdPurchaseJSONRequestBody
}

type PostItemsItemIdPurchaseResponseObject interface {
	VisitPostItemsItemIdPurchaseResponse(w http.ResponseWriter) error
}

type PostItemsItemIdPurchase200JSONResponse struct {
	ItemId    *string `json:"itemId,omitempty"`
	Purchased *int    `json:"purchased,omitempty"`

	// Remaining Units left in stock after the purchase.
	Remaining *int `json:"remaining,omitempty"`
}

func (response PostItemsItemIdPurchase200JSONResponse) VisitPostItemsItemIdPurchaseResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type PostItemsItemIdPurchase400JSONResponse Error

func (response PostItemsItemIdPurchase400JSONResponse) VisitPostItemsItemIdPurchaseResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)

	return json.NewEncoder(w).Encode(response)
}

type PostItemsItemIdPurchase404JSONResponse Error

func (response PostItemsItemIdPurchase404JSONResponse) VisitPostItemsItemIdPurchaseResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(404)

	return json.NewEncoder(w).Encode(response)
}

type GetItemsItemIdStockRequestObject struct {
	ItemId string `json:"itemId"`
}

type GetItemsItemIdStockResponseObject interface {
	VisitGetItemsItemIdStockResponse(w http.ResponseWriter) error
}

type GetItemsItemIdStock200JSONResponse struct {
	ItemId *string `json:"itemId,omitempty"`

	// Remaining Units left in stock.
	Remaining *int `json:"remaining,omitempty"`
}

func (response GetItemsItemIdStock200JSONResponse) VisitGetItemsItemIdStockResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type GetItemsItemIdStock404JSONResponse Error

func (response GetItemsItemIdStock404JSONResponse) VisitGetItemsItemIdStockResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(404)

	return json.NewEncoder(w).Encode(response)
}

// StrictServerInterface represents all server handlers.
type StrictServerInterface interface {
	// List magical items
	// (GET /items)
	GetItems(ctx context.Context, request GetItemsRequestObject) (GetItemsResponseObject, error)
	// Get a magical item
	// (GET /items/{itemId})
	GetItemsItemId(ctx context.Context, request GetItemsItemIdRequestObject) (GetItemsItemIdResponseObject, error)
	// Buy a magical item
	// (POST /items/{itemId}/purchase)
	PostItemsItemIdPurchase(ctx context.Context, request PostItemsItemIdPurchaseRequestObject) (PostItemsItemIdPurchaseResponseObject, error)
	// Get remaining stock for an item
	// (GET /items/{itemId}/stock)
	GetItemsItemIdStock(ctx context.Context, request GetItemsItemIdStockRequestObject) (GetItemsItemIdStockResponseObject, error)
}

type (
	StrictHandlerFunc    = strictgin.StrictGinHandlerFunc
	StrictMiddlewareFunc = strictgin.StrictGinMiddlewareFunc
)

func NewStrictHandler(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares}
}

type strictHandler struct {
	ssi         StrictServerInterface
	middlewares []StrictMiddlewareFunc
}

// GetItems operation middleware
func (sh *strictHandler) GetItems(ctx *gin.Context, params GetItemsParams) {
	var request GetItemsRequestObject

	request.Params = params

	handler := func(ctx *gin.Context, request interface{}) (interface{}, error) {
		return sh.ssi.GetItems(ctx, request.(GetItemsRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetItems")
	}

	response, err := handler(ctx, request)

	if err != nil {
		ctx.Error(err)
		ctx.Status(http.StatusInternalServerError)
	} else if validResponse, ok := response.(GetItemsResponseObject); ok {
		if err := validResponse.VisitGetItemsResponse(ctx.Writer); err != nil {
			ctx.Error(err)
		}
	} else if response != nil {
		ctx.Error(fmt.Errorf("unexpected response type: %T", response))
	}
}

// GetItemsItemId operation middleware
func (sh *strictHandler) GetItemsItemId(ctx *gin.Context, itemId string) {
	var request GetItemsItemIdRequestObject

	request.ItemId = itemId

	handler := func(ctx *gin.Context, request interface{}) (interface{}, error) {
		return sh.ssi.GetItemsItemId(ctx, request.(GetItemsItemIdRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetItemsItemId")
	}

	response, err := handler(ctx, request)

	if err != nil {
		ctx.Error(err)
		ctx.Status(http.StatusInternalServerError)
	} else if validResponse, ok := response.(GetItemsItemIdResponseObject); ok {
		if err := validResponse.VisitGetItemsItemIdResponse(ctx.Writer); err != nil {
			ctx.Error(err)
		}
	} else if response != nil {
		ctx.Error(fmt.Errorf("unexpected response type: %T", response))
	}
}

// PostItemsItemIdPurchase operation middleware
func (sh *strictHandler) PostItemsItemIdPurchase(ctx *gin.Context, itemId string) {
	var request PostItemsItemIdPurchaseRequestObject

	request.ItemId = itemId

	var body PostItemsItemIdPurchaseJSONRequestBody
	if err := ctx.ShouldBindJSON(&body); err != nil {
		ctx.Status(http.StatusBadRequest)
		ctx.Error(err)
		return
	}
	request.Body = &body

	handler := func(ctx *gin.Context, request interface{}) (interface{}, error) {
		return sh.ssi.PostItemsItemIdPurchase(ctx, request.(PostItemsItemIdPurchaseRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "PostItemsItemIdPurchase")
	}

	response, err := handler(ctx, request)

	if err != nil {
		ctx.Error(err)
		ctx.Status(http.StatusInternalServerError)
	} else if validResponse, ok := response.(PostItemsItemIdPurchaseResponseObject); ok {
		if err := validResponse.VisitPostItemsItemIdPurchaseResponse(ctx.Writer); err != nil {
			ctx.Error(err)
		}
	} else if response != nil {
		ctx.Error(fmt.Errorf("unexpected response type: %T", response))
	}
}

// GetItemsItemIdStock operation middleware
func (sh *strictHandler) GetItemsItemIdStock(ctx *gin.Context, itemId string) {
	var request GetItemsItemIdStockRequestObject

	request.ItemId = itemId

	handler := func(ctx *gin.Context, request interface{}) (interface{}, error) {
		return sh.ssi.GetItemsItemIdStock(ctx, request.(GetItemsItemIdStockRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetItemsItemIdStock")
	}

	response, err := handler(ctx, request)

	if err != nil {
		ctx.Error(err)
		ctx.Status(http.StatusInternalServerError)
	} else if validResponse, ok := response.(GetItemsItemIdStockResponseObject); ok {
		if err := validResponse.VisitGetItemsItemIdStockResponse(ctx.Writer); err != nil {
			ctx.Error(err)
		}
	} else if response != nil {
		ctx.Error(fmt.Errorf("unexpected response type: %T", response))
	}
}